Improve performance of getRows function
For a small (3x3) matrix this produced a 162% increase in runs/second.

For a big (100x100) matrix, I could not measure the original getRows function (my browswer simply froze!). The new function is naturally slower but still achieves about 1400 runs/second on my machine.

The new function wraps the core Array.slice instead of unnecessarily converting to a list and back again.