Create new tail-recursion-based LU implementation with pivot
With this commit the bugs are finally ironed out. Presently the new LU function can be accessed through luPivotWrapper for the purposes of benchmarking.

It gives modest speed improvements for small matrices and overwhelming speed improvements for large matrices. There may well still be improvements to be found, e.g. in initially generating the list of columns in the matrix.